Refactor PageView and related slivers for adaptive cross-axis sizing
- Removed the deprecated SliverFittedPage class and integrated its functionality into SliverFillViewport.
- Introduced _AdaptivePageViewport to handle cross-axis fitting in PageView when wrapCrossAxis is true.
- Updated the viewport logic to ensure smooth transitions and proper sizing based on the currently visible child.
- Modified tests to validate the new behavior of PageView with wrapCrossAxis, ensuring height adapts to child sizes.
- Cleaned up imports and adjusted widget exports accordingly.
