+++ packages/flutter/lib/widgets.dart
@@ -147,7 +147,6 @@ export 'src/widgets/single_child_scroll_view.dart';
 export 'src/widgets/size_changed_layout_notifier.dart';
 export 'src/widgets/sliver.dart';
 export 'src/widgets/sliver_fill.dart';
-export 'src/widgets/sliver_fitted_page.dart';
 export 'src/widgets/sliver_floating_header.dart';
 export 'src/widgets/sliver_layout_builder.dart';
 export 'src/widgets/sliver_persistent_header.dart';
